+++ src/LDL/Framework/Base/Contracts/Type/ToArrayInterface.php
@@ -6,6 +6,8 @@ use LDL\Framework\Base\Exception\ToArrayException;
 
 interface ToArrayInterface extends LDLTypeInterface
 {
+    public const TO_ARRAY_TYPE_METHOD_NAME = 'toArray';
+
     /**
      * @throws ToArrayException
      * @param bool $useKeys