+++ src/LDL/Framework/Base/Contracts/Type/ToBooleanInterface.php
@@ -6,6 +6,8 @@ use LDL\Framework\Base\Exception\ToArrayException;
 
 interface ToBooleanInterface extends LDLTypeInterface
 {
+    public const TO_BOOLEAN_TYPE_METHOD_NAME = 'toBoolean';
+
     /**
      * @throws ToArrayException
      * @return bool