+++ src/LDL/Framework/Base/Contracts/Type/ToDoubleInterface.php
@@ -4,6 +4,8 @@ namespace LDL\Framework\Base\Contracts\Type;
 
 interface ToDoubleInterface extends LDLTypeInterface
 {
+    public const TO_DOUBLE_TYPE_METHOD_NAME = 'toDouble';
+
     /**
      * @return float
      */