+++ src/LDL/Framework/Base/Contracts/Type/ToScalarInterface.php
@@ -6,6 +6,8 @@ use LDL\Framework\Base\Exception\ToArrayException;
 
 interface ToScalarInterface extends LDLTypeInterface
 {
+    public const TO_SCALAR_TYPE_METHOD_NAME = 'toScalar';
+
     /**
      * @throws ToArrayException
      * @return string|int|bool|float